docs: split examples
